Make the sigle logout process less strict.
If the user session on the SP has already expired or is not found, the SP MAY
still send a LogoutResponse with a success status.

urn:oasis:names:tc:SAML:2.0:status:Requester indicates an error on
the requester's side. In this case, since the NameID does not match,
the Requester value should be used.

For a SP-initiated logout request, if the user session is not found,
a redirect to the logout landing page occurs.