Change default keyval state file location
Previously, the keyval state file was configured to be stored in the "conf.d"
directory. However, by default, the NGINX process does not have write access to
this directory, requiring users to either specify a different path or change
the directory permissions.

In this commit, we've changed the default path for the state file to
"/var/lib/nginx/state", which is more suitable for most Linux users and does
not require additional permission changes.